# api.yaml
openapi: 3.1.1
info:
  title: Doctree API
  description: API для иерархической системы хранения документов
  version: "2.0.0"
servers:
  - url: /api/v2
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateNodeRequest:
      type: object
      properties:
        title:
          type: string
        parentId:
          type: string
          format: uuid
    UpdateNodeRequest:
      type: object
      properties:
        title:
          type: string
        parentId:
          type: string
          format: uuid
...
    TreeNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        parentId:
          type: string
          format: uuid
          nullable: true
      required: [id, title]
    Tree:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        children:
          type: array
          items:
            $ref: "#/components/schemas/Tree"
security:
  - bearerAuth: []
paths:
...
  /trees/nodes:
      get:
        tags: [Tree]
        summary: Получить все узлы
        responses:
          "200":
            description: Список узлов
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    nodes:
                      type: array
                      items:
                        $ref: "#/components/schemas/TreeNode"
          "403":
            description: У пользователя нет прав
      post:
        tags: [Tree]
        summary: Создание узла дерева
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateNodeRequest"
        responses:
          "201":
            description: Узел создан
          "403":
            description: У пользователя нет прав администратора
          "400": 
            description: Некорректный запрос
  /trees/nodes/{id}:
    get:
      tags: [Tree]
      summary: Получить узел по id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Найденный узел
          content:
            application/json:
              schema:
                type: object
                properties:
                  node:
                    $ref: "#/components/schemas/TreeNode"
        "403":
            description: У пользователя нет прав
        "404":
          description: Узел не найден
    patch:
      tags: [Tree]
      summary: Обновить узел (название или родитель)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNodeRequest"
      responses:
        "200":
          description: Узел обновлен
        "400":
          description: Неправильный запрос
        "403":
          description: У пользователя нет прав администратора
        "404":
          description: Не найден узел
    delete:
      tags: [Tree]
      summary: Удалить узел
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Узел удален
        "400":
          description: Неправильный запрос
        "403":
          description: У пользователя нет прав администратора
        "404":
          description: Не найден узел
  /trees/root:
    get:
      tags: [Tree]
      summary: Получить все деревья
      responses:
        "200":
          description: Список деревьев
          content:
            application/json:
              schema:
                type: object
                properties:
                  trees:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tree"
        "403":
            description: У пользователя нет прав
    post:
      tags: [Tree]
      summary: Создать корень дерева
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
              required: [title]
      responses:
        "201":
          description: Корень создан
        "403":
          description: У пользователя нет прав администратора
        "400":
          description: Некорректные данные
  /trees/root/{id}:
    get:
      tags: [Tree]
      summary: Получить целое дерево по узлу внутри
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Дерево найдено
          content:
            application/json:
              schema:
                type: object
                properties:
                  tree:
                      $ref: "#/components/schemas/Tree"
        "400":
          description: Некорректный запрос
        "404": 
          description: Узел не найден
    delete:
      tags: [Tree]
      summary: Удалить дерево
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Дерево удалено
        "400":
          description: Некорректный запрос или узел не является корневым
        "403":
          description: У пользователя нет прав администратора
        "404":
          description: Узел не найден
  /trees/sub/{id}:
    get:
      tags: [Tree]
      summary: Получить поддерево
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Поддерево найдено
          content:
            application/json:
              schema:
                type: object
                properties:
                  tree:
                      $ref: "#/components/schemas/Tree"
        "400":
          description: Некорректный запрос
        "403":
            description: У пользователя нет прав
        "404": 
          description: Узел не найден
...